// sample-app/ui/src/types/filters.d.ts
// Copyright 2025 Google LLC
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     https://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

export interface LastWeekMonthFilters {
  timePeriod: "last-week" | "last-month" | "last-week-to-last-week"
  hourRange: [number, number]
  days: string[] // Days selection
}

export interface CustomTimeFilters {
  timePeriod: "custom"
  startDate: Date
  endDate: Date
  hourRange: [number, number]
  days: string[] // Days selection
}